+++ src/main/java/com/mybrainfficial/waltzProject/menuInfo/MenuInfoService.java
@@ -1,8 +1,6 @@
 package com.mybrainfficial.waltzProject.menuInfo;
 
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
@@ -12,7 +10,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 
-import com.mybrainfficial.waltzProject.boardInfo.BoardInfoVO;
 import com.mybrainfficial.waltzProject.userInfo.UserInfoVO;
 
 @Service
@@ -26,7 +23,7 @@ public class MenuInfoService {
 	
 	@PostConstruct
 	public void setMenuInfo() { 
-		this.menuInfo = menuInfoDAO.selectMenuInfoAll();
+		MenuInfoService.menuInfo = menuInfoDAO.selectMenuInfoAll();
 		logger.info("======================== init menuInfo in memory ================");
 	}
